backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
chromadb
python-multipart
python-dotenv
pypdf
